chore: reduce exported `structs` for public API, move units tests to private mod
