+++ tests/test_local_evaluation.rs
@@ -1,115 +1,13 @@
-use posthog_rs::{
-    FeatureFlag, FeatureFlagCondition, FeatureFlagFilters, FlagCache, FlagValue,
-    LocalEvaluationResponse, LocalEvaluator, Property,
-};
-use serde_json::json;
-use std::collections::HashMap;
-
 #[cfg(feature = "async-client")]
 use httpmock::prelude::*;
 #[cfg(feature = "async-client")]
 use posthog_rs::ClientOptionsBuilder;
 #[cfg(feature = "async-client")]
+use serde_json::json;
+#[cfg(feature = "async-client")]
+use std::collections::HashMap;
+#[cfg(feature = "async-client")]
 use std::time::Duration;
-
-#[test]
-fn test_local_evaluation_basic() {
-    // Create a cache and evaluator
-    let cache = FlagCache::new();
-    let evaluator = LocalEvaluator::new(cache.clone());
-
-    // Create a simple flag
-    let flag = FeatureFlag {
-        key: "test-flag".to_string(),
-        active: true,
-        filters: FeatureFlagFilters {
-            groups: vec![FeatureFlagCondition {
-                properties: vec![],
-                rollout_percentage: Some(100.0),
-                variant: None,
-            }],
-            multivariate: None,
-            payloads: HashMap::new(),
-        },
-    };
-
-    // Update cache with the flag
-    let response = LocalEvaluationResponse {
-        flags: vec![flag],
-        group_type_mapping: HashMap::new(),
-        cohorts: HashMap::new(),
-    };
-    cache.update(response);
-
-    // Test evaluation
-    let properties = HashMap::new();
-    let result = evaluator.evaluate_flag("test-flag", "user-123", &properties);
-
-    assert!(result.is_ok());
-    assert_eq!(result.unwrap(), Some(FlagValue::Boolean(true)));
-}
-
-#[test]
-fn test_local_evaluation_with_properties() {
-    let cache = FlagCache::new();
-    let evaluator = LocalEvaluator::new(cache.clone());
-
-    // Create a flag with property conditions
-    let flag = FeatureFlag {
-        key: "premium-feature".to_string(),
-        active: true,
-        filters: FeatureFlagFilters {
-            groups: vec![FeatureFlagCondition {
-                properties: vec![Property {
-                    key: "plan".to_string(),
-                    value: json!("premium"),
-                    operator: "exact".to_string(),
-                    property_type: None,
-                }],
-                rollout_percentage: Some(100.0),
-                variant: None,
-            }],
-            multivariate: None,
-            payloads: HashMap::new(),
-        },
-    };
-
-    // Update cache
-    let response = LocalEvaluationResponse {
-        flags: vec![flag],
-        group_type_mapping: HashMap::new(),
-        cohorts: HashMap::new(),
-    };
-    cache.update(response);
-
-    // Test with matching properties
-    let mut properties = HashMap::new();
-    properties.insert("plan".to_string(), json!("premium"));
-
-    let result = evaluator.evaluate_flag("premium-feature", "user-123", &properties);
-    assert!(result.is_ok());
-    assert_eq!(result.unwrap(), Some(FlagValue::Boolean(true)));
-
-    // Test with non-matching properties
-    let mut properties = HashMap::new();
-    properties.insert("plan".to_string(), json!("free"));
-
-    let result = evaluator.evaluate_flag("premium-feature", "user-456", &properties);
-    assert!(result.is_ok());
-    assert_eq!(result.unwrap(), Some(FlagValue::Boolean(false)));
-}
-
-#[test]
-fn test_local_evaluation_missing_flag() {
-    let cache = FlagCache::new();
-    let evaluator = LocalEvaluator::new(cache);
-
-    let properties = HashMap::new();
-    let result = evaluator.evaluate_flag("non-existent", "user-123", &properties);
-
-    assert!(result.is_ok());
-    assert_eq!(result.unwrap(), None);
-}
 
 #[cfg(feature = "async-client")]
 #[tokio::test]
@@ -196,53 +94,4 @@ async fn test_local_evaluation_with_mock_server() {
     // The actual result depends on whether the mock was hit and processed
 
     eval_mock.assert();
-}
-
-#[test]
-fn test_cache_operations() {
-    let cache = FlagCache::new();
-
-    // Create multiple flags
-    let flags = vec![
-        FeatureFlag {
-            key: "flag1".to_string(),
-            active: true,
-            filters: FeatureFlagFilters {
-                groups: vec![],
-                multivariate: None,
-                payloads: HashMap::new(),
-            },
-        },
-        FeatureFlag {
-            key: "flag2".to_string(),
-            active: true,
-            filters: FeatureFlagFilters {
-                groups: vec![],
-                multivariate: None,
-                payloads: HashMap::new(),
-            },
-        },
-    ];
-
-    let response = LocalEvaluationResponse {
-        flags: flags.clone(),
-        group_type_mapping: HashMap::new(),
-        cohorts: HashMap::new(),
-    };
-
-    cache.update(response);
-
-    // Test get_flag
-    assert!(cache.get_flag("flag1").is_some());
-    assert!(cache.get_flag("flag2").is_some());
-    assert!(cache.get_flag("flag3").is_none());
-
-    // Test get_all_flags
-    let all_flags = cache.get_all_flags();
-    assert_eq!(all_flags.len(), 2);
-
-    // Test clear
-    cache.clear();
-    assert!(cache.get_flag("flag1").is_none());
-    assert_eq!(cache.get_all_flags().len(), 0);
 }
